refactor screenshot generator
